binary golay code and cartesian product of finite sets
