+++ combinatorics/src/linear_codes/mod.rs
@@ -1,4 +1,5 @@
 pub mod binary_golay_code;
+pub mod binary_golay_code_old;
 pub mod hexacode;
-pub mod ternary_golay_code;
-pub mod tetracode;
+pub mod ternary_golay_code_old;
+pub mod tetracode_old;
