+++ combinatorics/src/linear_codes/ternary_golay_code_old.rs
