+++ combinatorics/src/linear_codes/tetracode_old.rs
